Improved CSV auto detect + euro decimal handling
Fixes #79
Changed CSV delimiter detection to only determine the delimiter based
on the column names (the first row). This is such a simple and obvious
way to do it, I don’t know why I didn’t think of it in the first place.
Also adds support for european formatted numbers (with a comma for the
decimal place)