Bug fixed object naming
At the wrapper level, i.e.,

- .bfPearsonCorrelation
- .bfKendallTau

I now include the result$stat entry to uniformly have the result. It
didn’t make sense for a .bfKendallTau object to have as argument
bfObject$r. When programming for Kendall’s tau, just retain tauObs in
the code, that is, in .bfCorrieKernelKendallTau etc etc