Corrected file extension when writing to OSF and read jasp files  without analyses
When files are saved to the OSF the file extension in the OSF file
input window was not updated to .jasp. This has been fixed.
It is now also possible to read a data file and save it directly to a
wasp file and read the wasp file back with errors.