BF with informative priors from One Sample t-test now functional
Added .generalSummaryTtestBFoneSample to compute the BF from a t-test
statistic to allow for informative priors in a one sample t-test. Was
used to replace the function used thus far. This function is a clone of
the existing function .generalTtestBF This seems to work properly now,
except for the plots that won’t change when a different prior is chosen.