Added prior widget to Summary Stats Independent Samples T-Test BF: mostly functional
Renamed .generalSummaryTtestBFOneSample to .generalSummaryTtestBF and
changed it to make it also compute the two sample case.
Added the prior widget to the interface and changed the R source to use
the .generalSummaryTtestBF function. Seems to work properly (except
that plots don’t seem to respond to changes in prior preferences).
Substantial code simplifications and restructuring for legibility.
TODO: Dienes priors.