Changes to be able to build a debian package without errors:
Removed minified backbone.js because lintian complains about it.
And having a minified javascript is maybe nice when we're on dialup but
the files are included so that doesnt matter either.
Removed minified jquery as well cause it wasn't used.

Added description to control, removed meta-package dependency (qt5-default), have to check if all requirements are included now.

For fileheaderrecord: when using std::cout one should also include iostream.

Extra clean option in JASP-Engine.pro to remove R/Library that is created in an extra buildstep because the source analysis of debian doesnt like it.

Removed execute permission-flag from RegressionLinear.json because it isnt actually a script or executable..

Changed references to jasp executable to be small-caps for linux/debian..