Edit structure of R code sending
R code sending from analysisforms is now simplified based on the suggestion by @boutinb: a direct reference to MainWindow is no longer necessary (see diff in `semsimpleform.cpp`).  A simple call to `runRScript` and a declaration of the return handler `rScriptDoneHandler` are now sufficient for calling R code from an analysis form.

Co-Authored-By: Joris Goosen <jorisgoosen@users.noreply.github.com>