This adds the first part of issue https://github.com/jasp-stats/INTERNAL-jasp/issues/411
It adds CustomContrastsTableView, a tableview that has "Add/Delete Rows" buttons and a "Reset" button.
If you specify columnName on the qml object it will show a single row of doubles with a column for each label/level that specified column has.
You can also specify a source, in that case it will use the first added term (so it only makes sense to connect it to a source that gives a single value (like an assignedVariablesList with "singleVariable: true"))
See TestModule -> Custom Contrasts Widget for an example.