Add multiple errors
Several controls may have error. Each one should have its own error
message.
Moreover if there is an error on a control hidden by a section, then
the section should get a red border.
And if an analysis has control with an error, it should not send the
options to the engine.