Changed equivalence help files (#4019)
* Updated prior posterior plot

X-axis effect size added

* Added density table

Added a table with the densities over the prior and posterior in the equivalence and non-equivalence interval

* Update grammar error in helpful

* Changed helpfiles

* Update JASP-Engine/JASP/R/equivalencebayesianpairedsamplesttest.R

Co-Authored-By: Don van den Bergh <donvdbergh@hotmail.com>

* Update JASP-Engine/JASP/R/equivalencebayesianonesamplettest.R

Co-Authored-By: Don van den Bergh <donvdbergh@hotmail.com>

* Update JASP-Engine/JASP/R/equivalencebayesianindependentsamplesttest.R

Co-Authored-By: Don van den Bergh <donvdbergh@hotmail.com>

* Inferiority testing

Incorporated Don's feedback and added inferiority testing to the Bayesian equivalence testing

* change table name

From density to mass

* helpfiles changed

* gettext()

* dependencies fixed

Added new dependencies

* Incorporated feedback don

Posterior and prior plot were not showing when the Bayes Factor was infinity. Now a plot displays with a warning.

* Changed helpfiles

* Inferiority testing

Incorporated Don's feedback and added inferiority testing to the Bayesian equivalence testing

* Incorporated feedback don

Posterior and prior plot were not showing when the Bayes Factor was infinity. Now a plot displays with a warning.

* New helpfiles

* Added frequentist inferiority testing

Added frequentist inferiority testing and added cohen's d CI

Co-authored-by: Don van den Bergh <donvdbergh@hotmail.com>