fixes: #864
The variable in "Number of trials" was still being passed resulting in the error. I dealt with the situatuation in the same way as Bruno did in (same way as in https://github.com/jasp-stats/jasp-desktop/pull/4142). I hide and remove the variable from "Number of trials" if family is different that "Binomial ("Aggregated"). It should hopefully remove the arrow as well when Bruno merges his changes to the AssignedVariablesList.