Fixes multiple errors within the Mixed Models analysis (#4163)
* fixes: #862

A fitted model was deleted after setting an error withing the summary table. However, next ioption refreshed the model and it wasn't reset to null since the summary table wasn't called - the other output options could've been accessed without a proper model.
Fixed by creating a state that checks whether the model contained error. Doesn't allow to proceed unless a proper model was fitted.

* only GLMM didn't allow "nominalText" as a fixed effects

* fixes: #863

Only for LMMs require emmeans a type of df calculation to be specifie. If a GLMM model with gaussian family and identity link function is fitted, emmeans treat it as an LMM...
The same error would've occured for trends too, fixed both.

* fixes: #864

The variable in "Number of trials" was still being passed resulting in the error. I dealt with the situatuation in the same way as Bruno did in (same way as in https://github.com/jasp-stats/jasp-desktop/pull/4142). I hide and remove the variable from "Number of trials" if family is different that "Binomial ("Aggregated"). It should hopefully remove the arrow as well when Bruno merges his changes to the AssignedVariablesList.

* fixes #875

If R-hat is Inf, the max function returns NaN and it fails the logical test. A similar problem could've happend with ESS, which is produces as NaN directly by rstan

* translating variable names within an error message

* fixes #878

* fever -> fewer

* fixes #881

* fixes #865

by adding a checks for collinearity

* fixing more typos (Frantisek's English sucks)

* fixes #895

afex_plot requires 'dv' to be supplied when using stanova model

* redoing 879 after rebase

* fixes 892

by increasing minumum boostrap samples to 100

* fixing 893 (and another Frantisek's spelling mistake, his grammar sucks vol2)

* one last renaming that was lost during the rebase

* machine tollerance

Co-authored-by: Don van den Bergh <donvdbergh@hotmail.com>

* changing the +2 to the correct value

* removing the "hasERROR" jasp state

* always eval cont.bayes (should help with #884 and similar)

* error handling for B(G)LMM fits

* adding special check for iterations vs warmup

* more reasobable min values for the B(G)LMM

Co-authored-by: Don van den Bergh <donvdbergh@hotmail.com>