Special Unicode characters are still not well displayed on Windows
Fixes jasp-stats/jasp-issues#851

On Windows the gettext R function is apparently depending on the locale
of the machine, and sometimes transforms for example a \u03C9 to a
<U+03C9> instead of a ω character. This is afterwards considered as an
HTML tag by the WebEngine (and stays invisible). After many attempts to
understand why the gettext does not work correctly on Windows
(sometimes), a workaround has been decided to be added: transform in
the Engine any strings containing such a string ‘<U+FFFF>’ to its UTF8
character.