Setting an interface font makes data disappears
. The main problem was that when the interface font is changed, it
should call (like for uiScale change) the viewPort change slot of the
DataSetView
. The DropDown settings in PrefsUI.qml were confusing: the value was
set to be the preferencesModel.interfaceFont, but onValueChanged signal
sets the preferencesModel.interfaceFont to value.
. The value should not be directly set in a DropDown, only the
currentIndex. If no currentIndex is known, then a new property is made
‘startValue’ that sets internally the right currentIndex.
. Moreover, when an item is being initialised, the properties are
changed from empty to the specified value. So onValueChanged is called
twice (empty and specified value), without real interaction with the
user. These calls should be avoided, so the property ‘initialized’ has
been added, and is set to true only when the control is fully
initialised, and can be used to avoid these extra onValueChanged calls.
. Remove obsolete isDirectModel property in ComboBox