Refactor the Source Type for QML Controls and more... (#4413)
* Reactor the SourceType and more

Fixes some problems with ComboBox
Refactor the SourceType so that it becomes an apart class.
Make the column names as source type.
Start making the QML independent of the MainWindow. For this, all
changes of the dataset must be connected with the columnNames source
type.

* Add Joris' requests.

* Last small changes

* More fixes & rewrite

Change column variable does not remove anymore the variable from an
assigned list.
Remove the termsReset signal and uses the Qt model signals
Centralize the use and update of the terms in ListModel at one place.