Use of VariablesForm inside a TabView does not work
This is due to the initialisation of dynamic controls. They should not
be setup immediately when they are created, but when all dynamic
controls are created: when there are some dependencies between some
controls, they can then refer to each other being sure that they are
already created.
The VariablesForm is set as a JASPControl, so that we can beter control
when its initialisation starts: this initialisation sets the
relationships between the VariablesList inside the form, and those
relationships must be set before the VariablesLists self start their
setup.