Fix synching issue
Fixes https://github.com/jasp-stats/jasp-issues/issues/1193
There were 2 issues:  during synchronising, the dataset emits signals
that the data is changed, but the changes in the form should be handled
only by the datasetChanged signal in ColumnModel. All other signal
before should be ignored.
Als the Repeated Measures control did not handle well when some columns
are missing.
The RepeatedMeasuresFactorsList has been a bit reshaped to a
FactorLevelList so that it becomes a bit more generic.