Default Value with Binding issue
This issue was discovered when trying to solve
https://github.com/jasp-stats/jasp-issues/issues/1139

Open the ‘Default values with bindings’ analysis from the Test Module.
Assign some variables, uncheck the ‘check’ checkbox, set the `radio`
Radio Button to ‘three’ and set the Dropdown too ‘none’. Save the JASP
file, and load it again: the same values should be set.