Prophet analysis: adding seasonality may crash
In Prophet, if you change the name of a seasonality, and directly add a new one, it crashes.
This is caused by asking for a value in a QMap by directly calling the [] operator: if no value is found, it adds an empty entry. The QMap value method prevents this problem.