Update Position in the Tick table does not update always the label
In the plot editor, set manually the ticks. Changing the position should change automatically the label if they are equals. But is does not do that always.
The problem is that the data changed generates a rebuild of all cells of the TableView which is not needed. The DataSetView must handle also such a change correctly.