Handle better the constraints inside the Plot Editor
The constraints of the 'from' and 'to' values in the ticks settings are reinforced so that the 'from' value is always smaller than the 'to' value (with exception when the steps is negative: this is a strange default behaviour of the Raincloud plots).
Also when setting manually the ticks, it is not anymore possible to set the value of position smaller than its left value and bigger than its right value.

When the ticks are changed (either by specifying the sequence or setting them manually), the manual Lower & Upper limits (in Advanced) are changed also automatically as long as these limits are not previously changed manually (in fact as long as the upper & lower limits correspond to the 'from' and 'to' values).
The limit type (based on data, ticks or set manually) is not changed anymore.
This solves https://github.com/jasp-stats/jasp-test-release/issues/1502
and part of https://github.com/jasp-stats/jasp-test-release/issues/1385