rSource provokes running engine endlessly
This bug is accidentally revealed by the rSource, but it is a more general bug: when a value of a control is changed, it emits a signal so that the analysis must be rerun. This signal is sometimes blocked when some block code may change many values (initializing of analysis for example), and it is unblocked at the end of the block. When unblocked the signal was always emitted even if the signal was not called during the block code. But this signal must be emitted only if the signal was called during the block code.