The ASRM Mania Scale JASP file not found in Data Library
The name of the JASP file (and CSV file) did not correspond with the name is the json file.
Also by checking all Data Library JASP file, I noticed that when opening the BAIN SEM JASP file, the analysis is automatically refreshed. This is solved by a small easy fix in the TextArea handling.