Boost::filesystem is the one failing me! 🙄
For now, I leave it aside, it's quite easy to install Boost and tell CMake to find it. That's in fact my preferred method but we are moving toward independent build, so, when I figure out what's up with Boost, it's quite easy to just plug it in. For now, the `directory_iterator` seems to be culprit.