Experimenting with Modules/ path and installation. This is promising!
I have not yet fixed the paths but with these adjustments, I can get a clean build, and install everything from inside the `CMAKE_BINARY_DIR`. This mean, at the start of the build (once), I copy the `R.frameworks` inside the build folder as well, so we will have a similar debug and bundle builds and our paths are not altered that we have to take care of them later. This also leaves the original `R.framework` intact. This is important because, as we are starting to mess with it and installing `jags`, etc., we want to have it clean for clean builds.

→ I need to read a bit about the bundle packaging and relative paths and portable installation. CMake has a lot of tools to make a package "relocatable", as well as fixing rpath, and executable_path automatically and systematically.