The value of a control initilaized by a file should not change via binding
Fixes https://github.com/jasp-stats/jasp-test-release/issues/1760

If a value of a control is set with a binding in QML (for example a DoubleField is set per default by the number of variables selected by the user), this value can be changed by the user self to another value. If this value is saved in a JASP file, when reading the file, JASP must prevent the binding to overwrite the value.
This mechanism exists already via the _resetBindingValue JASPControl function. But in a case of an upgrade, an option that did not exist in a previous version, may have a default value set via a binding. In this particular case, the value should not be reset.