Trying to address the mysterious failing built on GitHub Action
It's a funny situation where GitHub env cannot sign the libraries, but not in a consistent way. For some reason, the command gets stuck and it often only get stuck during Build phase and not the configuration phase. So, I put a timeout on the signing command to make sure that it recalls it if it fails. Also, I'm adding a new flag to be skip the signing of some of third-party libraries during the build. So, if `SIGN_AT_BUILD_TIME` is set to OFF/FALSE only first party libraries like RInside, and Rcpp are getting signed _which are essential for R.framework to work properly during build._