Better CRYPT_KEY
Previously if you set `CRYPT_KEY` in `cmake` options it would just be ignored because `ENVIRONMENT_CRYPTKEY` wasn't set...
And the cryptkey in the environment was actually called `CRYPTKEY`so let's use that.

With the new logic it will use whatever you pass on to cmake first, then from `CRYPTKEY` in environment and lastly the default value as a fallback.
This way it also cannot fail which is nice and friendly.