Fixes to Modules.cmake
we do not need jaspTools at all in JASP.

Also, not setting `.libPaths(...)` means it will check for the normal R-library on your system as well..
So here locally for me building JASP I didnt get `jaspBase` because the check considered it fulfilled, jaspBase was already installed.
But then when running JASP (at least in the Install folder or in the resulting actual install of it) jaspBase was missing and no results showed up...
On a clean system this isnt a problem but I rarely work on a clean system and for people using jaspTools this will be the same.
So this