When Tab name changes, the bound value should be first updated
Fixes https://github.com/jasp-stats/jasp-test-release/issues/1822

When the name of a tab is changed, the nameChangedHandler of the ComponentsListBase is called, which calls the changeTerm method of its model. But as the name of a tab is also the key value of the bound value, this value must be first changed in the bound value so that the other control values of the tab can be found.