The dropdown value of old JASP files have sometimes not exactly the right value
Fixes https://github.com/jasp-stats/jasp-test-release/issues/1836

Due to old bugs, the value set in a dropdown might be set with the label and not with its value. This could be solved by adding entries in the upgrade.qml, but this can be quite cumbersome, since the problem can have occurred in many older JASP versions.
We can also argue, that if a value in a JASP file is not found in the possible value list of a DropDown, then before throwing an error message, we can try to find the value in the labels of the DropDown.