Row controls depending on a source should be disconnected when not used
Fixes https://github.com/jasp-stats/jasp-test-release/issues/1786

The MixedModels model uses a  ComponentsList with VariablesList having a source from another VariablesList. As the ComponentsList caches its controls, it must care that if a control has a source, changes in this source should not affect this control: the signals should be disconnected.