Cannot open the Column Filter when filter icon is visible
Fixes https://github.com/jasp-stats/jasp-issues/issues/1661

The problem was that the width of the LoadingIndicator of the column (in DataTableView) was not set to 0 when invisible.
The MouseArea that opens the Labels and Filters should also contain the Filter icon.
Also set the columnType as context variable: seems more logical.