Add combineTerms property ion soiurce
With the combineTerms source, it is now possible to combine the variables in different ways (cross, 2-way, 3-way, etc)
Also add the maxTermsWidth property in all list QML Widget, so that we can know what is the maximum length (depending of the metrics) of the terms of the list.
This is needed for @JohnnyDoorn 's new feature.