Integer Field shows wrong error
Fixes https://github.com/jasp-stats/INTERNAL-jasp/issues/2076

When an IntegerField (or a DoubleFiled) has a constraint (min or max for example), the control may show sometimes a red border as if it has wrong value, even if the value is OK.
This is due to a wrong fix made some time ago that activate this red border for controls that have wrong default value. This was done by assigning a width and a color to the focus indicator when the control was still initializing. This is wrong since the control get then wrong default values for its width and color border.
The check whether the control has wrong default value must be dome just after iti is initialized.