Admin maximum to engine count
Implements https://github.com/jasp-stats/jasp-issues/issues/1693

This can be used by providing a modified JASP config file to the users.
In it a value can be provided for the field `maxEngineCountAdmin`
and if >0 will be used to limit the amount of engines a user can spawn.
This can be helpful in constrained environment like citrix or some mainframe-client setup

If someone wants to use this feature and isn't quite sure how to do so, feel free to ask me in the above issue.