Allow Allowed & Suggested columns to be changed dynamically
The _setAllowedVariables function must be connected to the allowedColumnsChanged & suggestedColumnsChanged signals.
Moreover _setAllowedVariables should not mix up the _allowColumns & _suggestedColumns property: these properties should not be set in this function, only _variableTypesAllowed & _suggestedColumnsIcons should be set: _variableTypesAllowed is used to check which column is really allowed, and _suggestedColumnsIcons is used to set the icons in the VariablesList.