Order of variables with components must be kept
Fixes https://github.com/jasp-stats/INTERNAL-jasp/issues/2167

Problem is that the component values are stored in a map (RowControlsValues), and this map is used to set the option (in _getTableValueOption).
To fix it, the terms of the Variables List is also sent to this function, and the order of the terms is used to generate the option.