Remove Wrapper in wrapper function
Add in Descriptions.qml boolean hasWrappers, and for each analysis, the boolean hasWrapper and functionInternal properties.
If hasWrappers is true for the module, then all analyses have hasWrapper set to true.
If functionInternal is not set, then it takes the name of function with as suffix 'Internal' if hasWrapper is true.
When an analysis is called, the functionInternal is called: so this work for analysis having a wrapper or not. No upgrade is needed, since the name of the analysis stays the same.
In R Syntax the R code generated does not use the 'Wrapper' sufix anymore