Upgrader should not throw error if not a developer module
The errors were not send to the form: this was due to a wrong overriding of the upgradeMsgsForOption method.
An error message box should pop up only if the module is a Developer Module.
Make a difference between error and warning: If the issue comes from the fact that an option is not found in the JASP file, this is just a warning and should not be displayed in the form, only log it.