GITHUB_PAT should not disappear when navigating with tab key
Solves https://github.com/jasp-stats/INTERNAL-jasp/issues/2193

The TextInout echoMode property set to TextInput.PasswordEchoOnEdit provokes this problem. Using the echoMode to TextInput.Password solves this problem

Also use the editingFinished in place of textChanged signal with the PrefsTextInput control: textChanged is emitted for each key stroke, editingFinished is emitter with enter/return key or when the focus has gone.