Controls without name should work correctly
If isBound is set to false, then a name should not be compulsory. But the control must be setup correctly,
A special case is also handled here: when in a TabView, a user changes a title of a Tab, and directly afterwards add a new Tab. This was handled in a really hacky way, that prevents controls without name to be handled correctly.