The R Syntax setting should be kept
Fixes https://github.com/jasp-stats/INTERNAL-jasp/issues/2239

If the R Syntax is displayed on one analysis, then it is automatically displayed on all analyses, and also when JASP is started again.
In the same way the 'Show All R Options' is kept.

Also the PreferencesModelBase dummy object is added in order to make QMLComponents independent of the rest of JASP.