Highlight the right column header in Analyses mode
In Analyses mode, if the Variable Window is open, the corresponding column header should be highlighted.
Also by clicking on another column in the data view, if the Variable Window is already open, the Variable Window should display the settings of the clicked column
Remove the obsolete feature that the Variable disappear when a scale variable is chosen.