Edit/New data Ribbon should be enabled at the right moment
Fixes also https://github.com/jasp-stats/jasp-test-release/issues/2361

If there is no data, the New Data icon should be enabled, but if an analysis that does not require data (like a distribution) is added, then the New Data should be disabled.
If you remove the analysis, the New Data should be enabled.
If you open a JASP file that has no data, the New Data icon should be disabled.