Column type switch can lead to crash
Fixes https://github.com/jasp-stats/jasp-test-release/issues/2362

The crash was due to the fact that _labelByValueMap was not updated when a label was changed.
Also this PR adds the automatic switch to from nominalText to nominal if all values becomes integer. This is needed because if a user adds accidentally a string value to a nominal (or ordinal) column, the column is automatically changed to nominalText. If the user removes the wrong value, the column stays to nominalText, and the user has no way to change it back to nominal. JASP should detect this scenario and change the column to the tight type.