Add copy/paste with header
If a copy is done from the column header, the column name is also copied in the clipboard, so that if the user paste it in Excel, then the whole column with its name is copied.
If a paste is done in a cell in JASP, then the clipboard is copied.
If a paste is done in a column header, then if the copy was done form a column header, then the whole column (with name, labels, description...) is copied.
If a cut is done in a column header, the whole column is removed.

Moreover it ensures now that the name of a column is unique.