Add dataSetInfo properties
Fixes https://github.com/jasp-stats/INTERNAL-jasp/issues/2413

The QML files use dataSetModel.rowCount() and mainWindow.dataAvailiable. There are 2 problems with this:
. First QML analysis files should not use objects that are not defined outside the QMLComponents library (which will become a submodule), since these QML files will be used outside JASP (in R Studio).
. Second dataSetModel.rowCount()  is a function, not a property, so it cannot signal its changes. So it should be made as a property.

For this, an object dataSetInfo is made available in the root context with 2 properties dataAvailable & rowCount. The provider (here the ColumnsModel) has to take care that the signals are well propagated.