Check valid computed column from old JASP file
Fixes https://github.com/jasp-stats/jasp-issues/issues/2428

Apparently in the JASP file of the issue, a computed column was made that made the engine crash. This computed column is still written in the metadata.json, but it is ignored in 0.17.3. When upgrading to 0.18.1, this computed column is added and makes the engine crash. So just ignore this computed column when upgrading.