EmptyValues updates should work (also for files from 0.18.1) (#5347)
* EmptyValues updates should also work for files from 0.18.1

Make sure changes are updated in database
Also prepare the upgradeDB function in databaseInterface for future changes

* Workspace emptyvalues must have default values

When loading a 0.18.1 JASP file, the workspace default values are empty.
They should be set to the preferences empty values as default (since there is no workspace empty values in 0.18.1).
In DataSet::setEmptyValuesJson, there was already an handling to overcome this issue, but as DataSet has no access to the PreferencesModel object, it was depending of setting the _defaultEmptyvalues property just after the DataSet was created. But as DataSet might be deleted and created several times, the _defaultEmptyvalues is not always set. So just make it a static property, so that it just has to be set once.

---------

Co-authored-by: boutinb <b.boutin@uva.nl>