Remove ugly code
The Initialization of the available variablesList was quite ugly: a long comment in the bindTo method of the AnalysisForm object tried to explain how this initialization was done, but with this PR, I try to do it in the 'right way'.
The available VariablesLists have no bound value (they don't give values for the analysis), but they must be initialized. For this, as the setInitialized is called even for these controls, just override this method in VaraibelListBase, so that:
. when it has an avalable list model, then it should get its values from its sources.
. when it has an assigned list model, handle the particular case when the addAvailableVariablesToAssigned prperty is true, and it is initialized with no value, then it should take the values of its available list model (which might have default values).