A computed column is not always computed when opening a JASP file
Make a JASP file with several analyses.
Add a computed column by using another column.
Save it, close it, and open it again: the computed column is not computed due to an engine crash.

This crash is due to the fact that when trying to compute the computed column, the ColumnEncoder in the Engine was not yet initilalized with the names of the columns.