apply all upgrades and also try it on meta
previously a bug could occur in a specific situation:
say a module QC has an upgrade from 0.18.1 to 0.18.2 ijn 0.18.2. This works!
now some time later in version 0.19 an upgrade is added going from 0.18.2 to 0.19.
Because of flawed logic this dropped the 0.18.1->2 upgrade...

Besides that Ive added logging of options between each upgrade step,
this should help module developers out and bugreports as well

Then there is an attempt added to also update the `.meta` in options,
this should work for renames etc, and I dont think it will break anything else?