Add data should stay possible even after having adding an analysis
If you start JASP without file, add an analysis from Summary Statistics or Distribution that does not need data, it still should be possible to add data, either by loading a data file, or by adding manually data.

Also the icon for 'Edit Data' in Light theme seems to be a disabled icon. So I replace the grey color to black.