Do not remove custom empty value when editing cell
If a cell has a custom empty value, in a Edit mode when clicking on this cell, this removes the custom empty value.
For example, add 1 as empty value for the facFive column. Click on a cell that has 1 as value (it is displayed in grey), and go to another cell: the value 1 is removed.