Generate properly the markdown help
The info property is changed into a QVariant, so that more information can be set.
. For controls: if it is a string, this is as before just the text of the info. It can be used as an object like this: { text: "....", label: "....", useControlType: true, isHeader: true}
In this way the developer can specify whether the label can have another value than the title, whether the type of the control should be displayed and whether the info should be displayed as an header.

For Form, the info can be also just a string but also an object:
{ top: "..",, bottom: "...", references: ["ref1, "ref2"], RPackages: ["pack1", "pack2"], examples: ["example1", "example2"] }