stop edit-item from randomly jumping around when you click outside of it to stop editing
before this change, when editing (for instance) the third element of contBinom and you click another cell, it randomly jumps to a wrong one entirely.
this is caused by the reusing of items due to turning caching back on.
But this code fixes the actual problem by not moving the item with the mousearea around right in the middle of the click.