Datafile timestamp not stored in JASP file
Fixes https://github.com/jasp-stats/INTERNAL-jasp/issues/2611

The datafile timestamp was not stored in the JASP file, so when loading a JASP file, the data were always compared with the datafile. As the labels are not correctly compared, it thinks that the file has changed (cf https://github.com/jasp-stats/INTERNAL-jasp/issues/2614).

Moreover there was a bug in the JASP files in 0.18.3 (and earlier): when reading the sqlite database, the description and the datafile path were exchanged.