Set focus on column name when adding a new column (#5611)
In Edit mode, when the user click on the + icon to add a new column, he must first fill in the name to create the column, so that afterwards he can make a computed column. So disable all other field, and set the focus on the name field, so that the user understands this is the first action to do.